MySQL hack to override internal classes
