+++ src/MySql.EntityFrameworkCore.Design.Tests/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+//[assembly: InternalsVisibleTo("Webrox.EntityFrameworkCore.MySql PublicKeyToken=eea2c1a5c0e850ca")]
