+++ src/MySql.EntityFrameworkCore.Design.Tests/Directory.Build.props
@@ -1,0 +1,12 @@
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+	<PropertyGroup>
+		<EFVersionPackage>7.0</EFVersionPackage>
+		<TargetFramework>net7.0</TargetFramework>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="$(EFVersionPackage)" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="$(EFVersionPackage)" />
+		<PackageReference Include="MySql.EntityFrameworkCore"  Version="$(EFVersionPackage)" />
+	</ItemGroup>
+</Project>
