+++ src/Webrox.EntityFrameworkCore.MySql/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("MySql.EntityFrameworkCore PublicKeyToken=c5687fc88969c44d")]
